Move color calculation to utils
